Sort UIDs before slicing the during the initial sync
The logic of chunked and incremental initial syncs relied on the IMAP
server sorting UIDs strictly ascending. If that is not given, we fetch
messages with a UID too high for the current chunk. The next chunk will
continue with only higher UIDs than previous, leaving out any UIDs that
were missing from the previous chunk(s). Eventually the initial sync
finishes, but the mailbox is incomplete.

Signed-off-by: Christoph Wurst <christoph@winzerhof-wurst.at>